Stop transparently acting on child datasets
If a method should be performed on a child dataset the caller should
explicitly do so. This should prevent cases where the caller isn't sure
what happens when a dataset method is called.
Even more so now that the snapshot_list for datasets with children is
different from the standard datasets.