Fix unstable snapshot interval
Fix #38 by always clamping dates to a sane value for the given period.
Previously the targetted desired date for a given retention period would
depend on the day the backup was performed. As a result it shifted each
day by one day causing unstable intervals in between.

So a weekly retention period interval could be vary between 6 to 8 days.
For example weekly backups are now clamped to sundays (hardcoded, #39)
and as a result the best matching snapshot no longer shifts day by day.

This uses dateutil to perform the weekday clamping.
Add Django 3.1 as a supported version in the setup requirements.