Fix case where there are no snapshots to match
This woul happen for Filesets with large retention values or for new
Filesets which did not have enough snapshots to satisfy the retention
period.

Sort Dummy storage snapshots (almost) the same as zfs.
The creation date is not available so manual snapshots are not sorted as
they would be in zfs.