Add way to explicitly close Storage/Dataset
The TemporaryDirectory cleanup of the DummyStorage was left to the
garbage collection which could choose inconvenient times to remove
old directories. Thanks to the recently added pytest-randomly for
bringing this to our attention as well as some other naive tests that
relied on pre-existing data for their tests to pass.