contrib: planb-swiftsync: Begin fix to update mtime without rewriting all
Alters the list iterator to add a fourth case:

- only in left file
- only in right file
- in both files
- in both, but left and right have (only) a different mtime  -- added

We can use this new case to make lists to handle updated timestamps when
the data and filesize itself hasn't changed.

Useful when data has migrated from one large object storage to another
and the X-Timestamp has changed because of it.