contrib: manual-zfssync.sh: Add both --init and --prune subcommands
For an initial run, you'll want --init to fetch the bulk of the data.
This time you want many snapshots, as fetching things can take time. If
it only fetched the oldest snapshot, it might have been removed by the
time it's time for the incremental backups.

For the next runs, prefer without --init to get only N snapshots, making
the newest fetches fair.