rsync: Add pre-flight check that checks for /var/lib/planb/do-not-run.d files
If there are files there, we raise an error and a new job will
automatically get scheduled later on.

This is useful for cases when a large local backup is being made on a
host. If we want that complete backup to be backed up using planb, we'll
want to wait until it is complete.

The local backup job has as task to keep a file in
/var/lib/planb/do-not-run.d for the duration of the local backup run.
Once it removes the file, planb will be able to resume.