contrib: planb-swiftsync: Always update mtime in backups
If the file in OpenStack Swift had been updated with the same contents,
planb-swiftsync would not update the mtime.

This meant that a scan through the filesystem would show different
mtimes from the times in the planb-swiftsync.cur. That was confusing
and triggered a search for newer files.

From now on, we update the mtime when the file is the same but the mtime
wasn't.