contrib: planb-objsync: Configurable timeouts + defaults
Add the `planb_timeout` configuration option with a default of 60
seconds for all clients. The 60s default is adopted from the boto3
client because the Swift client default is to never timeout.

Note that both clients will retry failed requests.

For the boto3 client the connect and read timeout can be set
separately. The Swift client does not support this because
keystoneclient does not while the requests backend it uses does.

https://botocore.amazonaws.com/v1/documentation/api/latest/reference/config.html
https://docs.openstack.org/python-swiftclient/newton/swiftclient.html#swiftclient.client.Connection
https://docs.openstack.org/python-keystoneclient/latest/api/keystoneclient.v3.html#module-keystoneclient.v3.client
https://requests.readthedocs.io/en/latest/user/advanced/#timeouts